password login, token rotation, logout route

+++ src/app/api/v1/auth/utils/tokenRefresh.js
@@ -1,11 +1,13 @@
 import authDbConnect from "@/lib/mongodb/authDbConnect";
 import { sessionModel } from "@/models/auth/Session";
-import { decrypt } from "@/lib/encryption/cryptoEncryption";
+import { userModel } from "@/models/auth/User";
 import updateToken from "@/services/auth/updateToken";
 import jwt from "jsonwebtoken";
 import generateToken from "@/lib/generateToken";
-import crypto from 'crypto'; 
+import bcrypt from "bcryptjs";
 import config from "../../../../../../config";
+import { setSession } from "@/lib/redis/helpers/session";
+ 
 
 export async function tokenRefresh({ token }) {
 if (!token?.accessToken || !token?.refreshToken) 
@@ -15,39 +17,45 @@ if (!token?.accessToken || !token?.refreshToken)
       const { sessionId } = jwt.decode(token.accessToken, config.accessTokenSecret)
       if (!sessionId) 
         return null
-      const       auth_db = await authDbConnect();
-      const       Session = sessionModel(auth_db)  
+      const auth_db = await authDbConnect();
+      const Session = sessionModel(auth_db);
+      const    User = userModel(auth_db);
       const session = await Session.findOne({ _id: sessionId })
                                     .select('+refreshToken +revoked +refreshTokenExpiry' ).lean();
-      if(!session )  return null;
+      const user = await User.findOne({ referenceId: token.sub})
+
+      
+
+      if(!session || !user)  return null;
 
       if(Date.now() > session.refreshTokenExpiry || session.revoked == true ){
           await Session.deleteOne({  _id: sessionId  });
           return null;
       }
 
-      const oldRefreshToken = await decrypt({  cipherText: session.refreshToken, 
-                                                  options: { secret: config.refreshTokenEncryptionKey } })
+      // token.refreshToken --> from user request
+      // session.refreshToken --> from database
+      const isRefreshTokenValid = await bcrypt.compare(token.refreshToken, session.refreshToken);
 
-      const isMatch = crypto.timingSafeEqual( Buffer.from(oldRefreshToken),
-                                              Buffer.from(token.refreshToken) );
-      if (!isMatch) return null;
+      if (!isRefreshTokenValid) return null;
 
       const { accessToken,
               refreshToken,
               tokenId,
               accessTokenExpiry,
-              refreshTokenExpiry,
-              refreshTokenCipherText } = await generateToken({     user: {        email: token.email || '',
+              refreshTokenExpiry } = await generateToken({     user: {        email: token.email || '',
                                                                                   phone: token.phone || '',
                                                                             referenceId: token.sub }, 
                                                                 sessionId });
-
-      await updateToken({ sessionId, 
-                                data: {          tokenId, 
-                                      accessTokenExpiry, 
-                                            refreshToken: refreshTokenCipherText, 
-                                      refreshTokenExpiry  }})
+      await Promise.all([ updateToken({        db: auth_db,
+                                        sessionId, 
+                                             data: {          tokenId, 
+                                                    accessTokenExpiry, 
+                                                         refreshToken, 
+                                                    refreshTokenExpiry  }}),
+                          setSession({ sessionId, tokenId,
+                                      payload: { sub: user.referenceId, role: user.role } })
+                        ])                                                                
 
       return { accessToken, accessTokenExpiry, refreshToken  }
   } catch (error) {
